pyproject.toml
[tool.poetry]
name = "pycusseg"
version = "0.1.0"
description = "Customer Segmentation"
authors = ["Lê Huỳnh Đức <lhduc94@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12,<3.13"
pyaml = "^25.5.0"
pandas = "^2.3.0"
scikit-learn = "^1.7.0"
yellowbrick = "^1.5"
openpyxl = "^3.1.5"
seaborn = "^0.13.2"
setuptools = "^80.9.0"
numpy = "1.26"
dython = "^0.7.9"

[tool.poetry.group.dev.dependencies]
notebook = "^7.4.3"

[build-system]
requires = ["poetry-core", "setuptools", "wheel"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.urls]

